+++ src/main/java/com/codecool/report/dao/jdbc/StatusDaoJdbc.java
@@ -17,7 +17,7 @@ public class StatusDaoJdbc implements StatusDao {
         try {
             String sql = "INSERT INTO plant (id, status_name VALUES (?, ?)";
             PreparedStatement statement = conn.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS);
-            statement.setLong(1, status.getId());
+            statement.setInt(1, status.getId());
             statement.setString(2, status.getStatusName().getStatus());
             statement.executeUpdate();
             ResultSet resultSet = statement.getGeneratedKeys();
